Move proto ts files into src
